add index for controlers
