+++ controlers/index.js
@@ -1,0 +1,15 @@
+const listContacts = require("./listContacts");
+const getById = require("./getById");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+const removeContact = require("./removeContact");
+const updateFavorite = require("./updateFavorite");
+
+module.exports = {
+  listContacts,
+  getById,
+  addContact,
+  updateContact,
+  removeContact,
+  updateFavorite,
+};
