+++ controlers/listContacts.js
@@ -1,0 +1,9 @@
+const { Contact } = require("../models/contact");
+const { ctrlWrapper } = require("../helpers");
+
+const listContacts = async (_, res) => {
+  const result = await Contact.find();
+  res.json(result);
+};
+
+module.exports = ctrlWrapper(listContacts);
